requirements.txt
streamlit==1.31.0
ultralytics==8.2.0
supabase==2.9.1
pillow==10.2.0
numpy==1.26.4
pandas==2.2.0
